+++ src/mock/std/io.rs
@@ -211,7 +211,7 @@ mod mock_io {
         fn write_vectored(&mut self, bufs: &[std::io::IoSlice<'_>]) -> std::io::Result<usize> {
             match crate::macro_api::eval::<WriteMock::write_vectored>(self, bufs, &mut ()) {
                 Evaluation::Evaluated(result) => result,
-                Evaluation::Skipped(_) => IoFallback(self).write_vectored(bufs),
+                Evaluation::Unmocked(_) => IoFallback(self).write_vectored(bufs),
             }
         }
 
@@ -222,7 +222,7 @@ mod mock_io {
         fn write_all(&mut self, buf: &[u8]) -> std::io::Result<()> {
             match crate::macro_api::eval::<WriteMock::write_all>(self, buf, &mut ()) {
                 Evaluation::Evaluated(result) => result,
-                Evaluation::Skipped(_) => IoFallback(self).write_all(buf),
+                Evaluation::Unmocked(_) => IoFallback(self).write_all(buf),
             }
         }